requirements.txt
python-dotenv
pydantic>=2
pydantic-settings

livekit-agents
livekit-plugins-openai
livekit-plugins-google
livekit-plugins-silero
livekit-plugins-noise-cancellation

langchain_community
duckduckgo-search
requests
mem0ai
